+++ examples/app.py
@@ -129,7 +129,7 @@ if __name__=="__main__":
     # r=BrowserStarletteHTTP( Page )
     # r=BrowserStarletteWS( Page )
     # r=BrowserHTTP( Page )
-    r=ChromeApp( Page )
+    r=BrowserTornadoHTTP( Page )
     # r=WebHTTP( Page )
     r.run()